requirements.txt
streamlit==1.37.1
numpy==1.26.4
pandas==2.2.2
scipy==1.12.0
plotly==5.22.0
pyyaml==6.0.2
